--- backend/README.md
@@ -1,15 +1,0 @@
-back
-
-# how to run
-1. copy credentials.json from slack
-2. run
-go run .
-, then go to the link provided.
-3. click `agree` then copy the code in the query.
-4. run 
-go run .
-again. then pick one of the attachments (better be a jpg image) and paste the id to `sample.id`
-5. run
-go run .
-again. you'll find your file in `sample.jpg`. (if you choose another kind of file, change the extension name)
-
